Chats WIP
Restructured chats to not sit in dashboard, to reduce length of dashboard.jsx. Set what the basic idea of how group chats and DM chats will look, so I don't forget and also so the team can see what direction we are heading. Oh, also in the last commit titled 'Chats Section Update', I forgot to mention that I added a route for the friends list page.